+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Mapbox welcomes participation and contributions from everyone.
 ## main
 
 * Add a new CustomLayer API to simplify manipulation of layers with custom rendering (aka "CustomLayerHost").
+* Fix memory leak in SwiftUI.
 
 ## 11.0.0-beta.4 - 20 September, 2023